+++ src/main/java/edu/ie3/datamodel/io/csv/FileNamingStrategy.java
@@ -10,6 +10,7 @@ import edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.csv.timeseries.LoadProfileTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.naming.EntityPersistenceNamingStrategy;
 import edu.ie3.datamodel.io.naming.HierarchicFileNamingStrategy;
+import edu.ie3.datamodel.io.source.TimeSeriesMappingSource;
 import edu.ie3.datamodel.models.UniqueEntity;
 import edu.ie3.datamodel.models.input.AssetInput;
 import edu.ie3.datamodel.models.input.AssetTypeInput;
@@ -21,7 +22,6 @@ import edu.ie3.datamodel.models.result.ResultEntity;
 import edu.ie3.datamodel.models.timeseries.TimeSeries;
 import edu.ie3.datamodel.models.timeseries.TimeSeriesEntry;
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
-import edu.ie3.datamodel.models.timeseries.mapping.TimeSeriesMapping;
 import edu.ie3.datamodel.models.timeseries.repetitive.LoadProfileInput;
 import edu.ie3.datamodel.models.value.*;
 import edu.ie3.util.StringUtils;
@@ -248,7 +248,8 @@ public class FileNamingStrategy {
       return getGraphicsInputFileName(cls.asSubclass(GraphicInput.class));
     if (OperatorInput.class.isAssignableFrom(cls))
       return getOperatorInputFileName(cls.asSubclass(OperatorInput.class));
-    if (TimeSeriesMapping.Entry.class.isAssignableFrom(cls)) return getTimeSeriesMappingFileName();
+    if (TimeSeriesMappingSource.MappingEntry.class.isAssignableFrom(cls))
+      return getTimeSeriesMappingFileName();
     logger.error("There is no naming strategy defined for {}", cls.getSimpleName());
     return Optional.empty();
   }