+++ src/test/groovy/edu/ie3/datamodel/io/csv/HierarchicFileNamingStrategyTest.groovy
@@ -5,7 +5,7 @@
  */
 package edu.ie3.datamodel.io.csv
 
-import edu.ie3.datamodel.io.naming.HierarchicFileNamingStrategy
+import edu.ie3.datamodel.io.source.TimeSeriesMappingSource
 import edu.ie3.datamodel.models.BdewLoadProfile
 import edu.ie3.datamodel.models.UniqueEntity
 import edu.ie3.datamodel.models.input.MeasurementUnitInput
@@ -36,7 +36,6 @@ import edu.ie3.datamodel.models.result.thermal.CylindricalStorageResult
 import edu.ie3.datamodel.models.result.thermal.ThermalHouseResult
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue
-import edu.ie3.datamodel.models.timeseries.mapping.TimeSeriesMapping
 import edu.ie3.datamodel.models.timeseries.repetitive.LoadProfileInput
 import edu.ie3.datamodel.models.timeseries.repetitive.RepetitiveTimeSeries
 import edu.ie3.datamodel.models.value.EnergyPriceValue
@@ -426,7 +425,7 @@ class HierarchicFileNamingStrategyTest extends Specification {
 		def strategy = new HierarchicFileNamingStrategy(defaultHierarchy)
 
 		when:
-		def res = strategy.getDirectoryPath(TimeSeriesMapping.Entry)
+		def res = strategy.getDirectoryPath(TimeSeriesMappingSource.MappingEntry)
 
 		then:
 		res.present
@@ -438,7 +437,7 @@ class HierarchicFileNamingStrategyTest extends Specification {
 		def strategy = new HierarchicFileNamingStrategy(defaultHierarchy)
 
 		when:
-		def res = strategy.getFilePath(TimeSeriesMapping.Entry)
+		def res = strategy.getFilePath(TimeSeriesMappingSource.MappingEntry)
 
 		then:
 		res.present
@@ -450,7 +449,7 @@ class HierarchicFileNamingStrategyTest extends Specification {
 		def strategy = new HierarchicFileNamingStrategy("prefix", "suffix", defaultHierarchy)
 
 		when:
-		def res = strategy.getFilePath(TimeSeriesMapping.Entry)
+		def res = strategy.getFilePath(TimeSeriesMappingSource.MappingEntry)
 
 		then:
 		res.present