+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvTimeSeriesSourceTest.groovy
@@ -5,15 +5,16 @@
  */
 package edu.ie3.datamodel.io.source.csv
 
-import edu.ie3.datamodel.io.naming.EntityPersistenceNamingStrategy
+import static edu.ie3.datamodel.models.StandardUnits.ENERGY_PRICE
 
-import static edu.ie3.datamodel.models.StandardUnits.*
-
+import edu.ie3.datamodel.exceptions.SourceException
+import edu.ie3.datamodel.io.connectors.CsvFileConnector
+import edu.ie3.datamodel.io.csv.FileNamingStrategy
+import edu.ie3.datamodel.io.csv.timeseries.ColumnScheme
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedSimpleValueFactory
 import edu.ie3.datamodel.io.source.IdCoordinateSource
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue
-import edu.ie3.datamodel.models.timeseries.mapping.TimeSeriesMapping
-import edu.ie3.datamodel.models.value.EnergyPriceValue
+import edu.ie3.datamodel.models.value.*
 import edu.ie3.util.TimeUtil
 import edu.ie3.util.geo.GeoUtils
 import org.locationtech.jts.geom.Coordinate
@@ -23,36 +24,19 @@ import tech.units.indriya.quantity.Quantities
 import java.time.ZoneId
 
 class CsvTimeSeriesSourceTest extends Specification implements CsvTestDataMeta {
-	def "The csv time series source is able to provide a valid time series mapping from files"() {
-		given:
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new EntityPersistenceNamingStrategy())
-		def expectedMapping = [
-			new TimeSeriesMapping.Entry(UUID.fromString("58167015-d760-4f90-8109-f2ebd94cda91"), UUID.fromString("b86e95b0-e579-4a80-a534-37c7a470a409"), UUID.fromString("9185b8c1-86ba-4a16-8dea-5ac898e8caa5")),
-			new TimeSeriesMapping.Entry(UUID.fromString("9a9ebfda-dc26-4a40-b9ca-25cd42f6cc3f"), UUID.fromString("c7ebcc6c-55fc-479b-aa6b-6fa82ccac6b8"), UUID.fromString("3fbfaa97-cff4-46d4-95ba-a95665e87c26")),
-			new TimeSeriesMapping.Entry(UUID.fromString("9c1c53ea-e575-41a2-a373-a8b2d3ed2c39"), UUID.fromString("90a96daa-012b-4fea-82dc-24ba7a7ab81c"), UUID.fromString("3fbfaa97-cff4-46d4-95ba-a95665e87c26"))
-		]
-
-		when:
-		def mappingEntries = source.mapping
-
-		then:
-		mappingEntries.size() == expectedMapping.size()
-
-		expectedMapping.stream().allMatch { mappingEntries.contains(it) }
-	}
 
 	def "The csv time series source is able to build time based values from simple data"() {
 		given:
 		def defaultCoordinate = GeoUtils.DEFAULT_GEOMETRY_FACTORY.createPoint(new Coordinate(7.4116482, 51.4843281))
 		def coordinateSource = Mock(IdCoordinateSource)
 		coordinateSource.getCoordinate(5) >> defaultCoordinate
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new EntityPersistenceNamingStrategy())
 		def factory = new TimeBasedSimpleValueFactory(EnergyPriceValue)
+		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), UUID.fromString("2fcb3e53-b94a-4b96-bea4-c469e499f1a1"), "its_c_2fcb3e53-b94a-4b96-bea4-c469e499f1a1", EnergyPriceValue, factory)
 		def time = TimeUtil.withDefaults.toZonedDateTime("2019-01-01 00:00:00")
 		def timeUtil = new TimeUtil(ZoneId.of("UTC"), Locale.GERMANY, "yyyy-MM-dd'T'HH:mm:ss[.S[S][S]]'Z'")
 		def fieldToValue = [
-			"uuid": "78ca078a-e6e9-4972-a58d-b2cadbc2df2c",
-			"time": timeUtil.toString(time),
+			"uuid" : "78ca078a-e6e9-4972-a58d-b2cadbc2df2c",
+			"time" : timeUtil.toString(time),
 			"price": "52.4"
 		]
 		def expected = new TimeBasedValue(
@@ -67,5 +51,38 @@ class CsvTimeSeriesSourceTest extends Specification implements CsvTestDataMeta {
 		then:
 		actual.present
 		actual.get() == expected
+	}
+
+	def "The factory method in csv time series source refuses to build time series with unsupported column type"() {
+		given:
+		def metaInformation = new CsvFileConnector.CsvIndividualTimeSeriesMetaInformation(UUID.fromString("8bc9120d-fb9b-4484-b4e3-0cdadf0feea9"), ColumnScheme.WEATHER, "its_weather_8bc9120d-fb9b-4484-b4e3-0cdadf0feea9")
+
+		when:
+		CsvTimeSeriesSource.getSource(";", timeSeriesFolderPath, fileNamingStrategy, metaInformation)
+
+		then:
+		def e = thrown(SourceException)
+		e.message == "Unsupported column scheme '" + ColumnScheme.WEATHER + "'."
+	}
+
+	def "The factory method in csv time series source builds a time series source for all supported column types"() {
+		given:
+		def metaInformation = new CsvFileConnector.CsvIndividualTimeSeriesMetaInformation(uuid, columnScheme, path)
+
+		when:
+		def actual = CsvTimeSeriesSource.getSource(";", timeSeriesFolderPath, fileNamingStrategy, metaInformation)
+
+		then:
+		actual.timeSeries.entries.size() == amountOfEntries
+		actual.timeSeries.entries[0].value.class == valueClass
+
+		where:
+		uuid                                                    | columnScheme                                | path                                           || amountOfEntries | valueClass
+		UUID.fromString("2fcb3e53-b94a-4b96-bea4-c469e499f1a1") | ColumnScheme.ENERGY_PRICE                   | "its_c_2fcb3e53-b94a-4b96-bea4-c469e499f1a1"   || 2               | EnergyPriceValue
+		UUID.fromString("c8fe6547-fd85-4fdf-a169-e4da6ce5c3d0") | ColumnScheme.HEAT_DEMAND                    | "its_h_c8fe6547-fd85-4fdf-a169-e4da6ce5c3d0"   || 2               | HeatDemandValue
+		UUID.fromString("9185b8c1-86ba-4a16-8dea-5ac898e8caa5") | ColumnScheme.ACTIVE_POWER                   | "its_p_9185b8c1-86ba-4a16-8dea-5ac898e8caa5"   || 2               | PValue
+		UUID.fromString("76c9d846-797c-4f07-b7ec-2245f679f5c7") | ColumnScheme.ACTIVE_POWER_AND_HEAT_DEMAND   | "its_ph_76c9d846-797c-4f07-b7ec-2245f679f5c7"  || 2               | HeatAndPValue
+		UUID.fromString("3fbfaa97-cff4-46d4-95ba-a95665e87c26") | ColumnScheme.APPARENT_POWER                 | "its_pq_3fbfaa97-cff4-46d4-95ba-a95665e87c26"  || 2               | SValue
+		UUID.fromString("46be1e57-e4ed-4ef7-95f1-b2b321cb2047") | ColumnScheme.APPARENT_POWER_AND_HEAT_DEMAND | "its_pqh_46be1e57-e4ed-4ef7-95f1-b2b321cb2047" || 2               | HeatAndSValue
 	}
 }