+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   BREAKING: Harmonized field naming for time information
 -   BREAKING: Properly applying snake case to result file names
 -   deprecated `TarballUtils`
+-   updated Indriya to version 2.1.2 to include fixes for serialization
+-   Reworking the time series source (one source per time series, distinct mapping source, factory pattern)
 
 ### Fixed
 - InfluxDbConnector now keeps session instead of creating a new one each call
@@ -33,6 +35,7 @@ and [#248](https://github.com/ie3-institute/PowerSystemDataModel/issues/248))
 - BREAKING: fix invalid application of solar irradiance / irradiation ([#266](https://github.com/ie3-institute/PowerSystemDataModel/issues/266))
 - BREAKING: deleted `IrradiationValue` as it is invalid and no longer required
 - added copy builder implementations for thermal input models `CylindricalStorageInput`, `ThermalBusInput` and `ThermalHouseInput`
+- the sample code for reading and writing model data from respectively to csv files documented in the Sphinx documentation is fixed
 
 ## [1.1.0] - 2020-09-15